requirements.txt
pandas==2.2.2
numpy==1.26.4
nltk==3.8.1
scikit-learn==1.4.2
Flask==3.0.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
mysql-connector-python==8.3.0
selenium==4.18.1